refactor: move DateRange interface to base-date-cell-strategy and update imports
