+++ src/components/calendar/services/date-cells/current-month-date-cell-strategy.ts
@@ -1,6 +1,5 @@
-import { DateRange } from '../../types'
 import { getLastDateOfMonth } from '../../utils/get-last-date-of-month'
-import { BaseDateCellStrategy } from './base-date-cell-strategy'
+import { BaseDateCellStrategy, DateRange } from './base-date-cell-strategy'
 
 export class CurrentMonthDateCellStrategy extends BaseDateCellStrategy {
   calculateRange(year: number, month: number): DateRange {
