--- src/components/calendar/types/date-range.ts
@@ -1,4 +1,0 @@
-export interface DateRange {
-  startDate: number
-  endDate: number
-}
