+++ src/components/calendar/types/index.ts
@@ -1,3 +1,2 @@
 export * from './date-cell-renderer'
 export * from './date-cell-def'
-export * from './date-range'
